backend/pyproject.toml
[project]
name = "backend"
version = "2.0.0"
description = "DysLearnia — Visual learning pipeline builder for dyslexic learners"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-ollama>=0.2.0",
    "langgraph>=1.1.2",
    "ollama>=0.6.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sentence-transformers>=5.3.0",
    "uvicorn[standard]>=0.41.0",
    # IBlock system — optional heavy deps
    "pillow>=10.0.0",
    "python-pptx>=0.6.23",
    "markdown>=3.5.0",
    "qdrant-client>=1.9.0",
    "neo4j>=5.0.0",
    "supabase>=2.0.0",
    "websockets>=12.0",
    "pydub>=0.25.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.28.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
